+++ pulse2percept/utils/__init__.py
@@ -7,6 +7,7 @@
     constants
     geometry
     array
+    animation
     images
     convolution
     optimize
@@ -17,6 +18,7 @@
 """
 from .base import (PrettyPrint, FreezeError, Frozen, Parametrized, Data,
                    bijective26_name, cached, gamma)
+from .animation import HTMLAnimation
 from .geometry import (cart2pol, pol2cart, delta_angle)
 from .array import is_strictly_increasing, radial_mask, sample, unique
 from .images import center_image, scale_image, shift_image, trim_image
@@ -43,6 +45,7 @@ __all__ = [
     'FreezeError',
     'Frozen',
     'gamma',
+    'HTMLAnimation',
     'is_strictly_increasing',
     'Parametrized',
     'parse_3d_orient',